~ fixed skills with a time multiplier lower than 1 (Polaris and Test skill) not getting the correct skillpoints, this would also fix skills with a non-exact multiplier (like 1.5 or whatever). Fixes #111 You'll need to re-train that skill (remove it from the database and give it again)

+++ src/eve-common/utils/EvEMath.cpp
@@ -14,7 +14,7 @@ double EvEMath::Units::MetersToLightYears(double meters) {
 }
 
 // skill Equations
-uint32 EvEMath::Skill::PointsAtLevel(uint8 level, uint8 rank)
+uint32 EvEMath::Skill::PointsAtLevel(uint8 level, float rank)
 {
     if (level > EvESkill::MAXSKILLLEVEL)
         level = EvESkill::MAXSKILLLEVEL;
